# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Training Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /events:
    get:
      operationId: listEvents
      tags: [events]
      summary: List events visible to the current user
      security:
        - cookieAuth: []
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            default: 1
        - in: query
          name: limit
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Paginated list of events
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedEvents"

    post:
      operationId: createEvent
      tags: [events]
      summary: Create an in-person event (training_lead+)
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEventRequest"
      responses:
        "201":
          description: Event created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /events/{id}:
    get:
      operationId: getEvent
      tags: [events]
      summary: Get event details with registrations and attendance
      security:
        - cookieAuth: []
      parameters:
        - $ref: "#/components/parameters/EventId"
      responses:
        "200":
          description: Event detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventDetailResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    put:
      operationId: updateEvent
      tags: [events]
      summary: Update an event (training_lead+)
      security:
        - cookieAuth: []
      parameters:
        - $ref: "#/components/parameters/EventId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateEventRequest"
      responses:
        "200":
          description: Updated event
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventResponse"

    delete:
      operationId: deleteEvent
      tags: [events]
      summary: Soft-delete an event (training_lead+)
      security:
        - cookieAuth: []
      parameters:
        - $ref: "#/components/parameters/EventId"
      responses:
        "200":
          description: Event deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /events/{id}/register:
    post:
      operationId: registerForEvent
      tags: [events]
      summary: Self-register for an event
      security:
        - cookieAuth: []
      parameters:
        - $ref: "#/components/parameters/EventId"
      responses:
        "201":
          description: Registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventRegistrationResponse"
        "409":
          description: Already registered or full capacity
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /events/{id}/assign:
    post:
      operationId: assignUserToEvent
      tags: [events]
      summary: Assign a user to an event (training_lead+)
      security:
        - cookieAuth: []
      parameters:
        - $ref: "#/components/parameters/EventId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [userId]
              properties:
                userId:
                  type: string
      responses:
        "201":
          description: User assigned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventRegistrationResponse"
        "409":
          description: Already registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /events/{id}/registrations/{userId}:
    delete:
      operationId: unregisterFromEvent
      tags: [events]
      summary: Unregister a user from an event
      security:
        - cookieAuth: []
      parameters:
        - $ref: "#/components/parameters/EventId"
        - in: path
          name: userId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Unregistered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /events/{id}/attendance-code:
    post:
      operationId: generateAttendanceCode
      tags: [events]
      summary: Generate a one-time attendance code (training_lead+)
      security:
        - cookieAuth: []
      parameters:
        - $ref: "#/components/parameters/EventId"
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                expiresInMinutes:
                  type: integer
                  default: 60
      responses:
        "200":
          description: Attendance code generated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AttendanceCodeResponse"
  /events/{id}/attend:
    post:
      operationId: submitAttendanceCode
      tags: [events]
      summary: Submit an attendance code to mark self as attended
      security:
        - cookieAuth: []
      parameters:
        - $ref: "#/components/parameters/EventId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [code]
              properties:
                code:
                  type: string
      responses:
        "200":
          description: Attendance recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "400":
          description: Invalid or expired code
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /events/{id}/attendance/{userId}/mark:
    post:
      operationId: manualMarkAttendance
      tags: [events]
      summary: Manually mark a user as attended/not attended (training_lead+)
      security:
        - cookieAuth: []
      parameters:
        - $ref: "#/components/parameters/EventId"
        - in: path
          name: userId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                attended:
                  type: boolean
                  default: true
      responses:
        "200":
          description: Attendance updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /events/{id}/attendance/bulk-mark:
    post:
      operationId: bulkMarkAttendance
      tags: [events]
      summary: Mark all registered users as attended (training_lead+)
      security:
        - cookieAuth: []
      parameters:
        - $ref: "#/components/parameters/EventId"
      responses:
        "200":
          description: Bulk attendance marked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BulkMarkAttendanceResponse"
  /events/{id}/assignments:
    get:
      operationId: listEventAssignments
      tags: [events]
      summary: List group assignments for an event (training_lead+)
      security:
        - cookieAuth: []
      parameters:
        - $ref: "#/components/parameters/EventId"
      responses:
        "200":
          description: List of group assignments
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventAssignmentListResponse"

    post:
      operationId: assignEventToGroup
      tags: [events]
      summary: Assign an event to a group (admin only)
      security:
        - cookieAuth: []
      parameters:
        - $ref: "#/components/parameters/EventId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssignToGroupRequest"
      responses:
        "201":
          description: Assignment created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventAssignmentResponse"
        "409":
          description: Already assigned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /events/{id}/assignments/{assignmentId}:
    delete:
      operationId: unassignEventFromGroup
      tags: [events]
      summary: Remove an event group assignment (admin only)
      security:
        - cookieAuth: []
      parameters:
        - $ref: "#/components/parameters/EventId"
        - in: path
          name: assignmentId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Assignment removed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  # ─── Completions ─────────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: auth_token
  parameters:
...
    EventId:
      in: path
      name: id
      required: true
      schema:
        type: string
      description: Event UUID
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
    SuccessResponse:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
...
    Event:
      type: object
      required: [id, title, startAt, endAt, isActive, createdAt, updatedAt]
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: string
          nullable: true
        location:
          type: string
          nullable: true
        startAt:
          type: string
          format: date-time
        endAt:
          type: string
          format: date-time
        estimatedDurationMinutes:
          type: integer
          nullable: true
        maxCapacity:
          type: integer
          nullable: true
        attendanceCode:
          type: string
          nullable: true
          description: Only visible to training leads and admins
        attendanceCodeExpiresAt:
          type: string
          format: date-time
          nullable: true
        isActive:
          type: boolean
        createdBy:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    EventRegistration:
      type: object
      required: [id, eventId, userId, registeredAt]
      properties:
        id:
          type: string
        eventId:
          type: string
        userId:
          type: string
        assignedBy:
          type: string
          nullable: true
        registeredAt:
          type: string
          format: date-time
    EventGroupAssignment:
      type: object
      required: [id, eventId, groupId, assignedAt]
      properties:
        id:
          type: string
        eventId:
          type: string
        groupId:
          type: string
        assignedBy:
          type: string
          nullable: true
        assignedAt:
          type: string
          format: date-time

    # ─── Completion schemas ───────────────────────────────────────────────────
...
    AssignToGroupRequest:
      type: object
      required: [groupId]
      properties:
        groupId:
          type: string
        dueDate:
          type: string
          format: date-time
...
    CreateEventRequest:
      type: object
      required: [title, startAt, endAt]
      properties:
        title:
          type: string
        description:
          type: string
        location:
          type: string
        startAt:
          type: string
          format: date-time
        endAt:
          type: string
          format: date-time
        estimatedDurationMinutes:
          type: integer
        maxCapacity:
          type: integer
    UpdateEventRequest:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        location:
          type: string
        startAt:
          type: string
          format: date-time
        endAt:
          type: string
          format: date-time
        estimatedDurationMinutes:
          type: integer
        maxCapacity:
          type: integer
        isActive:
          type: boolean

    # ─── Response wrappers ────────────────────────────────────────────────────
...
    EventResponse:
      type: object
      required: [event]
      properties:
        event:
          $ref: "#/components/schemas/Event"
    EventDetailResponse:
      type: object
      required: [event, registrations, attendance]
      properties:
        event:
          $ref: "#/components/schemas/Event"
        registrations:
          type: array
          items:
            $ref: "#/components/schemas/EventRegistration"
        attendance:
          type: array
          items:
            type: object
            additionalProperties: true
    PaginatedEvents:
      type: object
      required: [events, total, page, limit]
      properties:
        events:
          type: array
          items:
            $ref: "#/components/schemas/Event"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
    EventRegistrationResponse:
      type: object
      required: [registration]
      properties:
        registration:
          $ref: "#/components/schemas/EventRegistration"
    AttendanceCodeResponse:
      type: object
      required: [code, expiresAt]
      properties:
        code:
          type: string
        expiresAt:
          type: string
          format: date-time
    BulkMarkAttendanceResponse:
      type: object
      required: [success, marked]
      properties:
        success:
          type: boolean
        marked:
          type: integer
    EventAssignmentResponse:
      type: object
      required: [assignment]
      properties:
        assignment:
          $ref: "#/components/schemas/EventGroupAssignment"
    EventAssignmentListResponse:
      type: object
      required: [assignments]
      properties:
        assignments:
          type: array
          items:
            $ref: "#/components/schemas/EventGroupAssignment"
...