# lib/api-zod/src/generated/types/bulkMarkAttendanceResponse.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Training Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface BulkMarkAttendanceResponse {
  success: boolean;
  marked: number;
}