# lib/api-zod/src/generated/types/eventAssignmentListResponse.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Training Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { EventGroupAssignment } from './eventGroupAssignment';

export interface EventAssignmentListResponse {
  assignments: EventGroupAssignment[];
}